backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation and Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# AI Integration
openai==1.12.0

# Utilities
python-dotenv==1.0.1
boto3==1.34.34
redis==5.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
